Refactor shared parser behavior.
