+++ src/Core/Ast/Parser/PhpStanParser/PhpStanParser.php
@@ -4,22 +4,20 @@ declare(strict_types=1);
 
 namespace Qossmic\Deptrac\Core\Ast\Parser\PhpStanParser;
 
-use PhpParser\Node\Stmt\ClassLike;
+use PhpParser\Node;
 use PhpParser\NodeTraverser;
 use PHPStan\Parser\Parser;
 use PHPStan\Parser\ParserErrorsException;
 use Qossmic\Deptrac\Contract\Ast\CouldNotParseFileException;
-use Qossmic\Deptrac\Core\Ast\AstMap\ClassLike\ClassLikeReference;
 use Qossmic\Deptrac\Core\Ast\AstMap\File\FileReference;
 use Qossmic\Deptrac\Core\Ast\AstMap\File\FileReferenceBuilder;
+use Qossmic\Deptrac\Core\Ast\Parser\AbstractParser;
 use Qossmic\Deptrac\Core\Ast\Parser\Cache\AstFileReferenceCacheInterface;
 use Qossmic\Deptrac\Core\Ast\Parser\Extractors\ReferenceExtractorInterface;
-use Qossmic\Deptrac\Core\Ast\Parser\ParserInterface;
 
-class PhpStanParser implements ParserInterface
+class PhpStanParser extends AbstractParser
 {
     private Parser $parser;
-    private readonly NodeTraverser $traverser;
 
     /**
      * @param ReferenceExtractorInterface<\PhpParser\Node>[] $extractors
@@ -39,25 +37,28 @@ class PhpStanParser implements ParserInterface
             return $fileReference;
         }
 
+        $scopeFactory = $this->phpStanContainer->createScopeFactory();
+        $reflectionProvider = $this->phpStanContainer->createReflectionProvider();
+
+        $fileReferenceBuilder = FileReferenceBuilder::create($file);
+        $visitor = new FileReferenceVisitor($fileReferenceBuilder, $scopeFactory, $reflectionProvider, $file, ...$this->extractors);
+        $nodes = $this->loadNodesFromFile($file);
+        $this->traverser->addVisitor($visitor);
+        $this->traverser->traverse($nodes);
+        $this->traverser->removeVisitor($visitor);
+
+        return $fileReferenceBuilder->build();
+    }
+
+    protected function loadNodesFromFile(string $filepath): array
+    {
         try {
-            $scopeFactory = $this->phpStanContainer->createScopeFactory();
-            $reflectionProvider = $this->phpStanContainer->createReflectionProvider();
+            $nodes = $this->parser->parseFile($filepath);
 
-            $fileReferenceBuilder = FileReferenceBuilder::create($file);
-            $visitor = new FileReferenceVisitor($fileReferenceBuilder, $scopeFactory, $reflectionProvider, $file, ...$this->extractors);
-            $nodes = $this->parser->parseFile($file);
-            $this->traverser->addVisitor($visitor);
-            $this->traverser->traverse($nodes);
-            $this->traverser->removeVisitor($visitor);
-
-            return $fileReferenceBuilder->build();
+            /** @var array<Node> $nodes */
+            return $nodes;
         } catch (ParserErrorsException $exception) {
             throw CouldNotParseFileException::because($exception->getMessage(), $exception);
         }
-    }
-
-    public function getNodeForClassLikeReference(ClassLikeReference $classReference): ?ClassLike
-    {
-        return null;
     }
 }
